Updated spawn manager system
A new function of "SpawnRandomAnimal" has been added in place of instantiating a random animal prefab directly from player's input. Now once player presses the "S" key on the keyboard, the function is called in place to yield the same result of instantiating a new animal prefab within the scene.